+++ final/Foundation4/Running.cs
@@ -4,7 +4,7 @@ class Running : Activity
 {
   private double _distance;
 
-  public Running(string date, int minutes, double distance, string type) : base(date, minutes, type)
+  public Running(string date, double minutes, double distance, string type) : base(date, minutes, type)
   {
     _distance = distance;
   }
